UI Fixes: placeholder, CredentialsForm, ErrorMessage
