+++ client/src/components/ErrorMessage/ErrorMessage.js
@@ -3,11 +3,7 @@ import styles from "./ErrorMessage.module.css";
 import { PropTypes } from "prop-types";
 
 export const ErrorMessage = ({ errorMessage }) => {
-  return errorMessage === "" ? (
-    <></>
-  ) : (
-    <div className={styles.errorMessage}>{errorMessage}</div>
-  );
+  return <div className={styles.errorMessage}>{errorMessage}</div>;
 };
 
 ErrorMessage.propTypes = { errorMessage: PropTypes.string };
