--- client/src/pages/CredentialsPage/CredentialsPage.js
@@ -1,54 +1,0 @@
-import React from "react";
-import CredentialsInputBox from "../../components/CredentialsInputBox/CredentialsInputBox.js";
-import { ThemeText } from "../../components/ThemeText/ThemeText.js";
-import { CredentialsButton } from "../../components/CredentialsButton/CredentialsButton.js";
-import { DescriptiveTextButton } from "../../components/DescriptiveTextButton/DescriptiveTextButton.js";
-import { useNavigate } from "react-router-dom";
-import PropTypes from "prop-types";
-
-import styles from "./CredentialsPage.module.css";
-
-const CredentialsPage = ({
-  headerText,
-  subtitleText,
-  actionText,
-  nextPage,
-  nextPageDescription,
-  nextPageText,
-}) => {
-  const navigate = useNavigate();
-  const routeToNextPage = () => {
-    navigate(nextPage);
-  };
-
-  return (
-    <div className={styles.bg}>
-      <div className={styles.credentials_container}>
-        <div style={{ marginBottom: "3px" }}>
-          <ThemeText primary={true} text={headerText} />
-        </div>
-        <ThemeText primary={false} text={subtitleText} />
-        <CredentialsInputBox errorMessage="This is a super long error message that can be editted." />
-        <div>
-          <CredentialsButton text={actionText} />
-        </div>
-        <DescriptiveTextButton
-          desc={nextPageDescription}
-          nextPageText={" " + nextPageText}
-          onClick={routeToNextPage}
-        />
-      </div>
-    </div>
-  );
-};
-
-CredentialsPage.propTypes = {
-  headerText: PropTypes.string.isRequired,
-  subtitleText: PropTypes.string.isRequired,
-  actionText: PropTypes.string.isRequired,
-  nextPage: PropTypes.string.isRequired,
-  nextPageDescription: PropTypes.string.isRequired,
-  nextPageText: PropTypes.string.isRequired,
-};
-
-export default CredentialsPage;
