+++ client/src/pages/CredentialsPage/SignUpPage.js
@@ -1,16 +1,21 @@
 import React from "react";
-import CredentialsPage from "./CredentialsPage";
+import CredentialsForm from "../../components/CredentialsForm/CredentialsForm";
+import styles from "./CredentialsPage.module.css";
 
 const SignUpPage = () => {
   return (
-    <CredentialsPage
-      headerText="Sign Up"
-      subtitleText="Create an account to continue"
-      actionText="Join Us"
-      nextPage="/"
-      nextPageDescription="Already have an account?"
-      nextPageText="Sign In"
-    />
+    <div className={styles.bg}>
+      <div className={styles.credentials_container}>
+        <CredentialsForm
+          headerText="Sign Up"
+          subtitleText="Create an account to continue"
+          actionText="Join Us"
+          nextPage="/"
+          nextPageDescription="Already have an account?"
+          nextPageText="Sign In"
+        />
+      </div>
+    </div>
   );
 };
